+++ cnccode/controller_classes/CTCustomerReviewMeeting.inc.php
@@ -236,6 +236,7 @@ class CTCustomerReviewMeeting extends CTCNC
                 }
 
                 $lastReviewMeetingDateFormatted = null;
+                $lastReviewMeetingDate = null;
                 if ($dsCustomer->getValue(DBECustomer::lastReviewMeetingDate)) {
                     $lastReviewMeetingDate = DateTime::createFromFormat(
                         DATE_MYSQL_DATE,
@@ -491,10 +492,11 @@ class CTCustomerReviewMeeting extends CTCNC
                     'reviews'
                 );
                 $dsReviews = new DataSet($this);
+
                 $buActivity->getManagementReviewsInPeriod(
                     $dsSearchForm->getValue(BUCustomerReviewMeeting::searchFormCustomerID),
-                    $startDate,
-                    $endDate,
+                    $lastReviewMeetingDate ? $lastReviewMeetingDate : $startDate,
+                    new DateTime(),
                     $dsReviews
                 );